RuboCop configuration file spec.
